requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
Django==3.0.6
django-extensions==2.2.9
django-js-asset==1.2.2
django-mptt==0.11.0
django-telegrambot==1.0.1
future==0.18.2
gunicorn==20.0.4
psycopg2==2.8.5
pycparser==2.20
python-telegram-bot==12.7
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
